Truncate data on reset / re-use existing metadata instance
After analyzing some slow-running unit test runs across a few services,
it became apparent that this could be largely attributed to
creating/dropping the database in between each test case.

Namely, there is a large cost attributed with evaluating each
table/column from scratch between tests. Locally, this could table up to
half a second(!). On some repos with 500+ tests, this adds up.

This change should reduce the cost of resetting the database (after the
initial run) milliseconds.

+++ microcosm_postgres/operations.py
@@ -2,6 +2,7 @@
 Common database operations.
 
 """
+from sqlalchemy import MetaData
 from sqlalchemy.exc import ProgrammingError
 
 from microcosm_postgres.migrate import main
@@ -65,13 +66,37 @@ def drop_alembic_table(graph):
         return True
 
 
+# Cached database metadata instance
+_metadata = None
+
+
 def recreate_all(graph):
     """
-    Drop and add back all database tables.
+    Drop and add back all database tables, or reset all data associated with a database.
+    Intended mainly for testing, where a test database may either need to be re-initialized
+    or cleared out between tests
 
     """
-    drop_all(graph)
-    create_all(graph)
+
+    global _metadata
+    initialize = False
+
+    if _metadata is None:
+        _metadata = MetaData(bind=graph.postgres, reflect=True)
+        initialize = True
+
+    # First use of this function, re-create the database from scratch.
+    if initialize:
+        drop_all(graph)
+        create_all(graph)
+        return
+
+    # Otherwise, truncate all existing tables
+    connection = graph.postgres.connect()
+    transaction = connection.begin()
+    for table in reversed(_metadata.sorted_tables):
+        connection.execute(table.delete())
+    transaction.commit()
 
 
 def new_session(graph, expire_on_commit=False):
